backend/requirements.txt
alembic==1.5.4
aniso8601==8.1.1
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
PyJWT==2.0.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
